Get closer to working scrolling
This version uses .drawImage() to draw on the canvas using a byte array rather than drawing it pixel by pixel. However, this produces blurry results. Either there's a way to do it using the canvas API and CSS or I'm going to have to reimplement the canvas transformation myself to so that it doesn't blur.
A bad case of the typos